Fixed a bug where we would use the form macros defined in plone.z3cform instead of our own.
Firstly, the registration was totally off, being for="BrowserLayer",
which is nonsense.  Secondly, that wouldn't work because it has to be
registered for a different layer.  Which is the CMF default layer
here, because Plone doesn't have one.

Then, for some unclear reason, we can't override the other
registration successfully without either 1) our own view class or 2) a
template definition in the class.

svn path=/plone.app.z3cform/trunk/; revision=21843