Translate fieldset labels correctly.
Note that the label is usually an i18n message, say with
label_fieldset as message id and Fieldset as default.

To translate the visible title on the tab, it is enough to use the
label with tal:content.  We should not use the default here, as this
default is not the message id that is translated.

For the html id and class we do not want translation, but we want a
string that is readable and not likely to change.  So we try three
possibilities.  First the '__name__' attribute of the group
(fieldset), then the default of the label, then the id of the label.
That last may be a string like 'label_my_fieldset', which is ugly, so
we prefer one of the others.