Add metal slot for inserting stuff below fields
It's currently not possible to insert any html by overriding the template on a form and using a fill-slot to insert something between the fields and the action buttons in a form. There's formtop and formbottom, but no belowfields. 

Example use case: we want to provide an extra snipped on a controlpanel form from plone.app.registry.browser.controlpanel, which generates everything for us, but it is almost impossible to add an extra text without resorting to inserting custom fields/etc.

Suggestions for a better name than 'belowfields' are appreciated.